requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
docopt==0.6.2
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-speech==2.19.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
Js2Py==0.74
oauthlib==3.2.2
packaging==23.1
pipwin==0.5.2
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
